pyproject.toml
[project]
name = "personal-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "a2a-sdk>=0.2.8",
    "arxiv-mcp-server>=0.1.0",
    "dotenv>=0.9.9",
    "fastapi>=0.115.12",
    "fastapi-mcp>=0.3.4",
    "fastmcp>=2.8.1",
    "google-adk>=1.3.0",
]
